requirements.txt
fastapi[all]
uvicorn
python-decouple
requests
pytest
httpx
